rename root folder
